+++ src/components/Cards/ButtonCard/index.tsx
@@ -1,15 +1,14 @@
 import { useCallback, useMemo } from "react";
 import styled from "@emotion/styled";
-import type { HassEntity } from "home-assistant-js-websocket";
 import { lowerCase, startCase } from "lodash";
-import type { EntityToServices, ServiceData } from "@typings";
-import {
-  useEntity,
-  useIconByDomain,
-  useIcon,
-  useApi,
-  useIconByEntity,
-} from "@hooks";
+import type {
+  DomainService,
+  ExtractDomain,
+  ServiceData,
+  HassEntityWithApi,
+  AllDomains,
+} from "@typings";
+import { useEntity, useIconByDomain, useIcon, useIconByEntity } from "@hooks";
 import { Ripples } from "../../Shared/Ripple";
 import { computeDomain } from "@utils/computeDomain";
 
@@ -23,7 +22,7 @@ export const StyledButtonCard = styled.button`
   display: flex;
   flex-direction: column;
   align-items: stretch;
-  justify-content: center;
+  justify-content: space-between;
   cursor: pointer;
   background-color: var(--ha-primary-background);
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
@@ -113,6 +112,7 @@ const LayoutBetween = styled.div`
   justify-content: space-between;
   flex-direction: row;
   margin-bottom: 20px;
+  gap: 10px;
 `;
 
 const LayoutRow = styled.div`
@@ -120,150 +120,158 @@ const LayoutRow = styled.div`
   align-items: center;
   justify-content: flex-start;
   flex-direction: row;
-  margin-bottom: 20px;
 `;
 
 const Title = styled.div`
   color: var(--ha-secondary-color);
   font-size: 0.7rem;
+  margin: 2px 0;
 `;
 
 const Description = styled.div`
   color: var(--ha-primary-color);
   font-size: 0.8rem;
-  padding-left: 6px;
 `;
-export interface BaseProps {
+export interface ButtonCardProps<
+  E extends `${AllDomains}.${string}`,
+  S extends DomainService<ExtractDomain<E>>
+> extends Omit<React.ComponentProps<"button">, "title" | "onClick"> {
   /** Optional icon param, this is automatically retrieved by the "domain" name if provided, or can be overwritten with a custom value  */
   icon?: string | null;
   /** the css color value of the icon */
   iconColor?: string | null;
-  /** the css color value of the background of the icon */
-  iconBackgroundColor?: string | null;
   /** By default, the title is retrieved from the domain name, or you can specify a manual title */
   title?: string | null;
   /** By default, the description is retrieved from the friendly name of the entity, or you can specify a manual description */
   description?: string | null;
-  /** optional classname to provide */
-  className?: string;
-}
-
-export interface ServiceProps<
-  E extends string,
-  S extends keyof EntityToServices<E>
-> extends BaseProps {
   /** The service name, eg "toggle, turnOn ..." */
-  service: S;
+  service?: S;
   /** The data to pass to the service */
-  serviceData?: ServiceData<E, S>;
+  serviceData?: ServiceData<ExtractDomain<E>, S>;
   /** The name of your entity */
-  entity: E;
-  /** the onClick handler is called when the button is pressed after the api service is called automatically.  */
-  onClick?: (entity: HassEntity) => void;
-}
-export interface ButtonCardProps<
-  E extends string,
-  S extends keyof EntityToServices<E>
-> extends Partial<ServiceProps<E, S>> {
+  entity?: E;
+  /** The onClick handler is called when the button is pressed, the first argument will be entity object with api methods if entity is provided  */
+  onClick?: (entity: HassEntityWithApi<ExtractDomain<E>>) => void;
   /** Optional active param, By default this is updated via home assistant */
   active?: boolean;
-  /** The onClick handler is called when the button is pressed  */
-  onClick?: () => void;
+  /** the layout of the button card, this changes slightly, just preferences really */
+  layout?: "default" | "slim";
 }
 
-function ServiceButton<E extends string, S extends keyof EntityToServices<E>>({
+export function ButtonCard<
+  E extends `${AllDomains}.${string}`,
+  S extends DomainService<ExtractDomain<E>>
+>({
   service,
-  serviceData,
   entity: _entity,
-  title: _title,
-  description: _description,
+  iconColor,
   icon: _icon,
+  active,
+  serviceData,
   onClick,
-}: ServiceProps<E, S>) {
-  const domain = computeDomain(_entity);
-  const api = useApi(domain);
-  const entity = useEntity(_entity);
-  const icon = useIconByDomain(domain);
-  const entityIcon = useIconByEntity(_entity);
-  const on = entity.state === "on";
-
-  const inputIcon = useIcon(_icon || null);
+  description: _description,
+  title: _title,
+  layout,
+  ...rest
+}: ButtonCardProps<E, S>): JSX.Element {
+  const domain = _entity ? computeDomain(_entity) : null;
+  const entity = useEntity(_entity || "number.non_existent", {
+    returnNullIfNotFound: true,
+  });
+  const icon = typeof _icon === "string" ? _icon : null;
+  const domainIcon = useIconByDomain(domain === null ? "unknown" : domain, {
+    color: iconColor || undefined,
+  });
+  const entityIcon = useIconByEntity(_entity || "number.non_existent", {
+    color: iconColor || undefined,
+  });
+  const isDefaultLayout = layout === "default" || layout === undefined;
+  const on = entity ? entity.state !== "off" : active || false;
+  // const { title, description, onClick, active, icon, ...others } = rest;
+  const iconElement = useIcon(icon, {
+    color: iconColor || undefined,
+  });
   const useApiHandler = useCallback(() => {
-    // @ts-expect-error - at this point we don't actually know the service
     // so we can expect it to throw errors however the parent level ts validation will catch invalid params.
-    api[service](entity.entity_id, serviceData);
-    if (typeof onClick === "function") onClick(entity);
-  }, [api, service, serviceData, entity, onClick]);
+    if (typeof service === "string" && entity) {
+      // @ts-expect-error - we don't actually know the service at this level
+      const caller = entity.api[service];
+      caller(serviceData);
+    }
+    if (typeof onClick === "function")
+      onClick(entity as HassEntityWithApi<ExtractDomain<E>>);
+  }, [service, entity, serviceData, onClick]);
+  // use the input description if provided, else use the friendly name if available, else entity name, else null
   const description = useMemo(() => {
     return _description === null
       ? null
-      : _description || entity?.attributes.friendly_name || entity.entity_id;
+      : _description ||
+          entity?.attributes.friendly_name ||
+          entity?.entity_id ||
+          null;
   }, [_description, entity]);
+  // use the input title if provided, else use the domain if available, else null
   const title = useMemo(
-    () => _title || startCase(lowerCase(domain)),
+    () => _title || (domain !== null ? startCase(lowerCase(domain)) : null),
     [_title, domain]
   );
-
-  return (
-    <StyledButtonCard onClick={useApiHandler}>
-      <LayoutBetween>
-        <Fab
-          brightness={(on && entity.custom.brightness) || "brightness(100%)"}
-          rgbaColor={
-            (on && entity.custom.rgbaColor) || "rgba(255,255,255,0.15)"
-          }
-          rgbColor={(on && entity.custom.rgbColor) || "white"}
-        >
-          {inputIcon || entityIcon || icon}
-        </Fab>
-        <Description>{description}</Description>
-      </LayoutBetween>
-      <Title>
-        {title} {entity.state} - {entity.custom.relativeTime}
-      </Title>
-    </StyledButtonCard>
-  );
-}
-/** The ButtonCard is a simple to use component to make it easy to control and visualize the state of a device. */
-export function ButtonCard<
-  E extends string,
-  S extends keyof EntityToServices<E>
->({
-  service,
-  entity,
-  iconColor,
-  iconBackgroundColor,
-  ...rest
-}: ButtonCardProps<E, S>) {
-  const isServiceButton = useMemo(() => {
-    return [service, entity].every((value) => typeof value !== "undefined");
-  }, [service, entity]);
-  const { title, description, onClick, active, icon, ...others } = rest;
-  const iconElement = useIcon(icon || null);
   return (
     <Ripples borderRadius="1rem">
-      {isServiceButton ? (
-        <ServiceButton {...rest} service={service!} entity={entity!} />
-      ) : (
-        <StyledButtonCard {...others} onClick={onClick}>
-          <LayoutBetween>
-            <Fab
-              brightness="brightness(100%)"
-              rgbaColor={iconBackgroundColor || "rgba(255,255,255,0.35)"}
-              rgbColor={iconColor || "rgb(255,255,255)"}
-            >
-              {iconElement}
-            </Fab>
-            <Toggle active={active === true}>
+      <StyledButtonCard {...rest} onClick={useApiHandler}>
+        <LayoutBetween>
+          <Fab
+            brightness={(on && entity?.custom.brightness) || "brightness(100%)"}
+            rgbaColor={
+              entity
+                ? on
+                  ? entity.custom.rgbaColor
+                  : "rgba(255,255,255,0.15)"
+                : on
+                ? "var(--ha-primary-active)"
+                : "var(--ha-primary-inactive)"
+            }
+            rgbColor={
+              entity
+                ? on
+                  ? entity.custom.rgbColor
+                  : "white"
+                : on
+                ? "var(--ha-secondary-active)"
+                : "var(--ha-secondary-inactive)"
+            }
+          >
+            {iconElement || entityIcon || domainIcon}
+          </Fab>
+          {isDefaultLayout && (
+            <Toggle active={on}>
               <ToggleState />
             </Toggle>
-          </LayoutBetween>
-          <LayoutRow>
-            {title && <Title>{title}</Title>}
-            {description && <Description>{description}</Description>}
-          </LayoutRow>
-        </StyledButtonCard>
-      )}
+          )}
+          {!isDefaultLayout && <Description>{description}</Description>}
+        </LayoutBetween>
+        <LayoutRow>
+          {!isDefaultLayout && (
+            <Title>
+              {title}{" "}
+              {typeof active === "boolean"
+                ? active
+                  ? "- on"
+                  : "- off"
+                : entity
+                ? `- ${entity.state}`
+                : ""}
+              {entity ? ` - ${entity.custom.relativeTime}` : ""}
+            </Title>
+          )}
+          {isDefaultLayout && (
+            <Title>
+              {title && <Title>{title}</Title>}
+              {description && <Description>{description}</Description>}
+              {entity && <Title>Updated: {entity.custom.relativeTime}</Title>}
+            </Title>
+          )}
+        </LayoutRow>
+      </StyledButtonCard>
     </Ripples>
   );
 }