+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
 export { useHass } from "./useHass";
 export { useEntity } from "./useEntity";
-export { useApi } from "./useApi";
+export { useApi, createService } from "./useApi";
 export { useIconByDomain, useIcon, useIconByEntity } from "./useIcon";