+++ src/utils/computeDomain.ts
@@ -1,5 +1,9 @@
-import { DomainName } from "@typings";
+import { AllDomains, CamelToSnake } from "@typings";
 import { snakeCase } from "lodash";
 
-export const computeDomain = (entityId: string): DomainName =>
-  snakeCase(entityId.substr(0, entityId.indexOf("."))) as DomainName;
+export const computeDomain = <E extends `${AllDomains}.${string}`>(
+  entityId: E
+): CamelToSnake<AllDomains> =>
+  snakeCase(
+    entityId.substr(0, entityId.indexOf("."))
+  ) as CamelToSnake<AllDomains>;